requirements.txt
fastapi
uvicorn
httpx
websockets
pydantic
python-multipart
python-dotenv
pymongo
numpy
opencv-python
pillow
mediapipe
onnxruntime
scikit-learn
scipy
PyYAML
open3d
torch
torchvision
ultralytics
transformers
accelerate
qwen-vl-utils

